requirements.txt
anthropic>=0.42.0
simple-salesforce>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
flask>=3.0.0
gunicorn>=21.2.0
